Split palindrom.test file into three
