+++ test/specs/palindrom.test.js
@@ -238,391 +238,3 @@ describe("Palindrom", () => {
     });
   });
 });
-describe("Palindrom", () => {
-  describe("#error responses", () => {
-    beforeEach(() => {
-      moxios.install();
-    });
-    afterEach(() => {
-      moxios.uninstall();
-    });
-    it("should call onConnectionError on HTTP 400 response", function(done) {
-      const spy = sinon.spy();
-
-      moxios.stubRequest("http://localhost/testURL", {
-        status: 400,
-        headers: { contentType: "application/json" },
-        responseText: "Custom message"
-      });
-
-      let tempObject;
-      const that = this;
-
-      const palindrom = new Palindrom({
-        remoteUrl: "http://localhost/testURL",
-        onConnectionError: spy
-      });
-
-      /* onConnectionError should be called once now */
-      setTimeout(
-        () => {
-          assert(spy.calledOnce);
-          done();
-        },
-        5
-      );
-    });
-
-    it("should call onConnectionError on HTTP 599 response", function(done) {
-      const spy = sinon.spy();
-
-      moxios.stubRequest("http://localhost/testURL", {
-        status: 599,
-        headers: { contentType: "application/json" },
-        responseText: "Custom message"
-      });
-
-      let tempObject;
-      const that = this;
-
-      const palindrom = new Palindrom({
-        remoteUrl: "http://localhost/testURL",
-        onConnectionError: spy
-      });
-
-      /* onConnectionError should be called once now */
-      setTimeout(
-        () => {
-          assert(spy.calledOnce);
-          done();
-        },
-        5
-      );
-    });
-
-    it("should call onConnectionError on HTTP 400 response (patch)", function(done) {
-      const spy = sinon.spy();
-
-      moxios.stubRequest("http://localhost/testURL", {
-        status: 200,
-        headers: { contentType: "application/json" },
-        responseText: '{"hello": "world"}'
-      });
-
-      const palindrom = new Palindrom({
-        remoteUrl: "http://localhost/testURL",
-        onConnectionError: spy
-      });
-
-      moxios.withMock(() => {
-        setTimeout(
-          () => {
-            palindrom.obj.hello = "galaxy";
-            let request = moxios.requests.mostRecent();
-            request
-              .respondWith({
-                status: 400,
-                headers: { contentType: "application/json" },
-                response: "Custom message"
-              })
-              .then(() => {
-                /* onConnectionError should be called once now */
-                assert(spy.calledOnce);
-                done();
-              });
-          },
-          5
-        );
-      });
-    });
-    it("should call onConnectionError on HTTP 599 response (patch)", function(done) {
-      const spy = sinon.spy();
-
-      moxios.stubRequest("http://localhost/testURL", {
-        status: 200,
-        headers: { contentType: "application/json" },
-        responseText: '{"hello": "world"}'
-      });
-
-      const palindrom = new Palindrom({
-        remoteUrl: "http://localhost/testURL",
-        onConnectionError: spy
-      });
-
-      moxios.withMock(() => {
-        setTimeout(
-          () => {
-            palindrom.obj.hello = "galaxy";
-            let request = moxios.requests.mostRecent();
-            request
-              .respondWith({
-                status: 599,
-                headers: { contentType: "application/json" },
-                response: "Custom message"
-              })
-              .then(() => {
-                assert(spy.calledOnce);
-                done();
-              });
-          },
-          5
-        );
-      });
-    });
-  });
-});
-describe("Palindrom", () => {
-  describe("#IgnoreAdd", () => {
-    beforeEach(() => {
-      moxios.install();
-    });
-    afterEach(() => {
-      moxios.uninstall();
-    });
-    it("should not send add patch to an ignored property", function(done) {
-      const spy = sinon.spy();
-
-      moxios.stubRequest("http://localhost/testURL", {
-        status: 200,
-        headers: { contentType: "application/json" },
-        responseText: '{"hello": "world"}'
-      });
-
-      let tempObject;
-      const palindrom = new Palindrom({
-        remoteUrl: "http://localhost/testURL",
-        onConnectionError: spy,
-        callback: function(myObj) {
-          tempObject = myObj;
-        }
-      });
-
-      palindrom.ignoreAdd = /\/\$.+/;
-
-      /* before adding an ignored variable, we should have one ajax request */
-      setTimeout(
-        () => {
-          assert.equal(1, moxios.requests.count());
-
-          /* now we add an ignored variable */
-          tempObject.$privateProp = 1;
-        },
-        5
-      );
-
-      /* after adding an ignored variable, we should still have one ajax request */
-      setTimeout(
-        () => {
-          assert.equal(1, moxios.requests.count());
-
-          /* now added a NOT ignored variable */
-          tempObject.publicProb = 1;
-        },
-        10
-      );
-
-      /* after adding a NOT ignored variable, we should have TWO ajax requests */
-      setTimeout(
-        () => {
-          assert.equal(2, moxios.requests.count());
-          let request = moxios.requests.mostRecent();
-
-          /* and the mostRecent of them should be the following patch */
-          assert.equal(
-            '[{"op":"add","path":"/publicProb","value":1}]',
-            request.config.data
-          );
-          done();
-        },
-        15
-      );
-    });
-    it("should not send replace patch to an ignored property", function(done) {
-      const spy = sinon.spy();
-
-      moxios.stubRequest("http://localhost/testURL", {
-        status: 200,
-        headers: { contentType: "application/json" },
-        responseText: '{"hello": "world"}'
-      });
-
-      let tempObject;
-      const palindrom = new Palindrom({
-        remoteUrl: "http://localhost/testURL",
-        onConnectionError: spy,
-        callback: function(myObj) {
-          tempObject = myObj;
-        }
-      });
-
-      palindrom.ignoreAdd = /\/\$.+/;
-
-      /* before adding an ignored variable, we should have one ajax request */
-      setTimeout(
-        () => {
-          assert.equal(1, moxios.requests.count());
-
-          /* add an ignored variable */
-          tempObject.$privateProp = 1;
-        },
-        5
-      );
-
-      /* let's change the ignored variable */
-      setTimeout(
-        () => {
-          tempObject.$privateProp = 2;
-        },
-        10
-      );
-
-      /* after changing an ignored variable, we should still have one ajax request */
-      setTimeout(
-        () => {
-          assert.equal(1, moxios.requests.count());
-          done();
-        },
-        15
-      );
-    });
-
-    it("should not send replace patch to an ignored deep object", function(done) {
-      const spy = sinon.spy();
-
-      moxios.stubRequest("http://localhost/testURL", {
-        status: 200,
-        headers: { contentType: "application/json" },
-        responseText: '{"hello": 0}'
-      });
-
-      const palindrom = new Palindrom({
-        remoteUrl: "http://localhost/testURL",
-        onConnectionError: spy,
-        callback: function(tempObject) {
-          tempObject.publicProp = [1, 2, 3];
-
-          setTimeout(
-            () => {
-              /* initial request and `add` publicProb */
-              assert.equal(2, moxios.requests.count());
-
-              tempObject.$privateProp = [1, 2, 3];
-
-              /* we should have two requests, initial and publicProb add */
-              setTimeout(
-                () => {
-                  assert.equal(2, moxios.requests.count());
-                },
-                5
-              );
-
-              /* change ignored property deeply */
-              setTimeout(
-                () => {
-                  tempObject.$privateProp[1] = 32;
-                },
-                10
-              );
-
-              /* we should STILL have two requests, initial and publicProb add only */
-              setTimeout(
-                () => {
-                  assert.equal(2, moxios.requests.count());
-                  done();
-                },
-                15
-              );
-            },
-            5
-          );
-        }
-      });
-
-      palindrom.ignoreAdd = /\/\$.+/;
-    });
-
-    it("should not send any patch if all changes were ignored", function(done) {
-      const spy = sinon.spy();
-
-      moxios.stubRequest("http://localhost/testURL", {
-        status: 200,
-        headers: { contentType: "application/json" },
-        responseText: '{"hello": 0}'
-      });
-
-      let tempObject;
-      const palindrom = new Palindrom({
-        remoteUrl: "http://localhost/testURL",
-        onConnectionError: spy,
-        callback: function(myObj) {
-          tempObject = myObj;
-
-          tempObject.$privateProp = 1;
-        }
-      });
-
-      palindrom.ignoreAdd = /\/\$.+/;
-
-      /* we should have two requests, initial and publicProb add */
-      setTimeout(
-        () => {
-          assert.equal(1, moxios.requests.count());
-
-          tempObject.$privateProp = 22;
-        },
-        5
-      );
-
-      /* change ignored property deeply */
-      setTimeout(
-        () => {
-          assert.equal(1, moxios.requests.count());
-          done();
-        },
-        100
-      );
-    });
-
-    it("should not send a patch when added property is replaced", function(done) {
-      const spy = sinon.spy();
-
-      moxios.stubRequest("http://localhost/testURL", {
-        status: 200,
-        headers: { contentType: "application/json" },
-        responseText: '{"hello": 0}'
-      });
-
-      let tempObject;
-      const palindrom = new Palindrom({
-        remoteUrl: "http://localhost/testURL",
-        onConnectionError: spy,
-        callback: function(myObj) {
-          tempObject = myObj;
-        }
-      });
-
-      palindrom.ignoreAdd = /\/\$.+/;
-
-      /* we should have one request, initial connection */
-      setTimeout(
-        () => {
-          assert.equal(1, moxios.requests.count());
-
-          /* change ignored properties */
-          tempObject.$privateProp = 1;
-          tempObject.$privateProp = 2;
-        },
-        5
-      );
-
-      /*  we should still have one request, initial connection */
-      setTimeout(
-        () => {
-          assert.equal(1, moxios.requests.count());
-          done();
-        },
-        100
-      );
-    });
-  });
-});
